requirements/dev.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --no-emit-index-url --output-file=requirements/dev.txt requirements/dev.in
#
alabaster==0.7.16
    # via sphinx
attrs==23.2.0
    # via pytest-mypy
babel==2.14.0
    # via sphinx
beautifulsoup4==4.12.2
    # via furo
boto3-stubs==1.34.19
    # via -r requirements/dev.in
botocore-stubs==1.34.19
    # via
    #   -r requirements/dev.in
    #   boto3-stubs
build==1.0.3
    # via pip-tools
cachetools==5.3.2
    # via
    #   -c requirements/main.txt
    #   google-auth
    #   tox
certifi==2023.11.17
    # via
    #   -c requirements/main.txt
    #   requests
chardet==5.2.0
    # via tox
charset-normalizer==3.3.2
    # via
    #   -c requirements/main.txt
    #   requests
click==8.1.7
    # via
    #   -c requirements/main.txt
    #   pip-tools
cloud-storage-mocker==0.3.1
    # via -r requirements/dev.in
colorama==0.4.6
    # via tox
commonmark==0.9.1
    # via recommonmark
coverage[toml]==7.4.0
    # via
    #   coverage
    #   pytest-cov
distlib==0.3.8
    # via virtualenv
docutils==0.20.1
    # via
    #   recommonmark
    #   sphinx
filelock==3.13.1
    # via
    #   pytest-mypy
    #   tox
    #   virtualenv
flake8==7.0.0
    # via -r requirements/dev.in
furo==2023.9.10
    # via -r requirements/dev.in
google-api-core==2.15.0
    # via
    #   -c requirements/main.txt
    #   google-cloud-core
    #   google-cloud-storage
google-auth==2.26.2
    # via
    #   -c requirements/main.txt
    #   google-api-core
    #   google-auth-stubs
    #   google-cloud-core
    #   google-cloud-storage
google-auth-stubs==0.2.0
    # via -r requirements/dev.in
google-cloud-core==2.4.1
    # via
    #   -c requirements/main.txt
    #   google-cloud-storage
google-cloud-storage==2.14.0
    # via
    #   -c requirements/main.txt
    #   cloud-storage-mocker
google-crc32c==1.5.0
    # via
    #   -c requirements/main.txt
    #   google-cloud-storage
    #   google-resumable-media
google-resumable-media==2.7.0
    # via
    #   -c requirements/main.txt
    #   google-cloud-storage
googleapis-common-protos==1.62.0
    # via
    #   -c requirements/main.txt
    #   google-api-core
grpc-stubs==1.53.0.5
    # via google-auth-stubs
grpcio==1.60.0
    # via grpc-stubs
idna==3.6
    # via
    #   -c requirements/main.txt
    #   requests
    #   yarl
imagesize==1.4.1
    # via sphinx
iniconfig==2.0.0
    # via pytest
isort==5.13.2
    # via pytest-isort
jinja2==3.1.3
    # via
    #   -c requirements/main.txt
    #   sphinx
markupsafe==2.1.3
    # via
    #   -c requirements/main.txt
    #   jinja2
mccabe==0.7.0
    # via flake8
multidict==6.0.4
    # via yarl
mypy==1.8.0
    # via pytest-mypy
mypy-extensions==1.0.0
    # via mypy
packaging==23.2
    # via
    #   -c requirements/main.txt
    #   build
    #   pyproject-api
    #   pytest
    #   sphinx
    #   tox
pip-tools==7.3.0
    # via -r requirements/dev.in
platformdirs==4.1.0
    # via
    #   tox
    #   virtualenv
pluggy==1.3.0
    # via
    #   pytest
    #   tox
protobuf==4.25.2
    # via
    #   -c requirements/main.txt
    #   google-api-core
    #   googleapis-common-protos
pyasn1==0.5.1
    # via
    #   -c requirements/main.txt
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.3.0
    # via
    #   -c requirements/main.txt
    #   google-auth
pycodestyle==2.11.1
    # via flake8
pyflakes==3.2.0
    # via flake8
pygments==2.17.2
    # via
    #   furo
    #   sphinx
pyproject-api==1.6.1
    # via tox
pyproject-hooks==1.0.0
    # via build
pytest==7.4.4
    # via
    #   -r requirements/dev.in
    #   pytest-cov
    #   pytest-env
    #   pytest-isort
    #   pytest-mypy
    #   pytest-vcr
pytest-cov==4.1.0
    # via -r requirements/dev.in
pytest-env==1.1.3
    # via -r requirements/dev.in
pytest-isort==3.1.0
    # via -r requirements/dev.in
pytest-mypy==0.10.3
    # via -r requirements/dev.in
pytest-vcr==1.0.2
    # via -r requirements/dev.in
pytz==2023.3.post1
    # via -r requirements/dev.in
pyyaml==6.0.1
    # via
    #   -c requirements/main.txt
    #   vcrpy
recommonmark==0.7.1
    # via -r requirements/dev.in
requests==2.31.0
    # via
    #   -c requirements/main.txt
    #   google-api-core
    #   google-cloud-storage
    #   sphinx
rsa==4.9
    # via
    #   -c requirements/main.txt
    #   google-auth
snowballstemmer==2.2.0
    # via sphinx
soupsieve==2.5
    # via beautifulsoup4
sphinx==7.2.6
    # via
    #   -r requirements/dev.in
    #   furo
    #   recommonmark
    #   sphinx-autodoc-typehints
    #   sphinx-basic-ng
sphinx-autodoc-typehints==1.25.2
    # via -r requirements/dev.in
sphinx-basic-ng==1.0.0b2
    # via furo
sphinxcontrib-applehelp==1.0.8
    # via sphinx
sphinxcontrib-devhelp==1.0.6
    # via sphinx
sphinxcontrib-htmlhelp==2.0.5
    # via sphinx
sphinxcontrib-jsmath==1.0.1
    # via sphinx
sphinxcontrib-qthelp==1.0.7
    # via sphinx
sphinxcontrib-serializinghtml==1.1.10
    # via sphinx
tox==4.12.0
    # via -r requirements/dev.in
types-awscrt==0.20.0
    # via botocore-stubs
types-cryptography==3.3.23.2
    # via types-jwt
types-jwt==0.1.3
    # via -r requirements/dev.in
types-python-dateutil==2.8.19.20240106
    # via -r requirements/dev.in
types-pytz==2023.3.1.1
    # via -r requirements/dev.in
types-pyyaml==6.0.12.12
    # via -r requirements/dev.in
types-requests==2.31.0.20240106
    # via google-auth-stubs
types-s3transfer==0.10.0
    # via boto3-stubs
typing-extensions==4.9.0
    # via
    #   -c requirements/main.txt
    #   boto3-stubs
    #   mypy
urllib3==2.0.7
    # via
    #   -c requirements/main.txt
    #   requests
    #   types-requests
vcrpy==5.1.0
    # via pytest-vcr
virtualenv==20.25.0
    # via tox
wheel==0.42.0
    # via pip-tools
wrapt==1.16.0
    # via vcrpy
yarl==1.9.4
    # via vcrpy

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
